+++ macos-leo/app/Components/MainApps/Safari/components/IFrameBottom.tsx
@@ -1,0 +1,18 @@
+import { IFrameBottomProps } from "../Type/ManeFrameProps";
+
+  
+function IFrameBottom({ url }: IFrameBottomProps) {
+    return (
+      <div className="w-full h-full">
+        <iframe 
+          src={url} 
+          width="100%" 
+          height="600px" 
+          allowFullScreen
+          title="Website Preview"
+        />
+      </div>
+    );
+  }
+  
+  export default IFrameBottom;